More work on chunking; should be just about complete. Need to test.
git-svn-id: svn://okws.org/okws2/devel/1.1@2069 7d287422-6eea-da11-b677-00123f2a783a

+++ libpub/pub.h
@@ -1087,7 +1087,7 @@ public:
   // for chunked file xfer in pub2
   void init_xdr_opaque ();
   ssize_t len () const;
-  ssize_t get_chunk (off_t offset, char *buf, size_t capacity) const;
+  ssize_t get_chunk (size_t offset, char *buf, size_t capacity) const;
   void get_xdr_hash (xpubhash_t *x) const { _xdr_opaque_hash.to_xdr (x); }
 protected:
   str _xdr_opaque;
